bot/requirements.txt
python-dotenv~=1.0.1
psycopg2~=2.9.9
paramiko~=3.4.0
python-telegram-bot==13.7